requirements.txt
asgiref==3.8.1
contourpy==1.3.1
cycler==0.12.1
Django==4.2.19
django-cors-headers==3.14.0
djangorestframework==3.15.2
fonttools==4.56.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.3
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
pyparsing==3.2.1
python-dateutil==2.9.0.post0
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
